Redesign the control scheme and document it.
Add an option to toggle between "standard" and "enhanced" controls
which mainly changes how the automatic thermostat is configured
using the keypad.  The keypad behavior with the enhanced controls is
modeless so the buttons always do the same thing regardless of whether
the thermostat is on.  The thermostat can be overridden just by pressing
a button to change the state of the fan and then it can be re-engaged
by pressing the auto button again.  The user guide explains everything.

Rewrote keypad decoding to support chorded keypresses and to immediately
dispatch actions when buttons are held long enough instead of waiting
for them to be released first.

Add rain sensor.
Add factory reset.
Add wifi toggle.
Add power-on restore behavior and toggle.
Support all features of the 6 key, 4 key, and IR controls.
Add placeholders for audible feedback.