Support brushed fan motors on the MCT8316Z.
This is a bit of a hack because it requires resetting the motor lock
timeout periodically but it's enough to compare the behavior of the
motors.  The SPI variant of this chip allows the lock timeout to be
disabled.  Too bad it isn't I2C as that would be easier to work with.

For the factory motor, connect the white wire to motor phase U and
the black wire to motor phase V with a JST VH-3 connector.  Also
connect the hall V sensor to ground (wire pin 3 to pin 1 of the
JST XH-5 connector).